Rename FloatinActionButton to FloatingActionButton
