+++ src/app/components/FloatingActionButton/FloatingActionButton.stories.tsx
